(TextEditor) Remove all things Jodit related
