+++ containers/mosaic/src/components/Field/FormFieldTextEditor/index.ts
@@ -1,9 +1,7 @@
-export { default } from "./FormFieldTextEditor";
+export { FormFieldTextEditor as default } from "./FormFieldTextEditor";
 export * from "./FormFieldTextEditor";
-export * from "./FormFieldTextEditorTipTap";
 export * from "./FormFieldTextEditorTypes";
 
-export * from "./FormFieldTextEditorTipTapFieldType";
 export * from "./Extensions";
 export * from "./NodeForm";
 export * from "./Toolbar";
