+++ containers/sb-8/stories/components/Field/FormFieldTextEditor/FormFieldTextEditor.mdx
@@ -1,43 +1,95 @@
 import { Canvas, Meta } from '@storybook/blocks';
 import * as stories from './FormFieldTextEditor.stories';
 
-<Meta of={stories} name="Docs - Jodit (Legacy)" />
+<Meta of={stories} name="Docs - Tiptap" />
 
-## Text Editor - Jodit (Legacy)
+## Text Editor - Tiptap
+The integration of this field into the form system is not complete. It is being included alongside the existing Jodit editor to ensure correct alignment between the two fields before replacing it completely. To use it, import the `FormFieldTextEditorTipTapFieldType` and pass that as the field type:
+
+```ts
+import { FormFieldTextEditorTipTapFieldType } from "@simpleview/sv-mosaic/components";
+
+const fields: FieldDef = [
+	{
+		name: "My Text Editor",
+		type: FormFieldTextEditorTipTapFieldType,
+		// ...
+	}
+]
+
+```
+
 This field implements the [**FieldDef**](#generic-field-props-fielddef) interface.
 
 - The text editor is a text area with added capabilities for use in various publishers, allowing the users to format their input in a text area.
-- This implementation is based on the [Jodit WYSIWYG Editor](https://www.npmjs.com/package/jodit) package.
-- [**Playground**](/?path=/story/formfields-formfieldtexteditor--playground)
-- Data: `string` - Value of the input typed.
+- This is an implementation using the [Tiptap](https://www.npmjs.com/package/@tiptap/react) package.
+- [**Playground**](/story/formfields-formfieldtexteditor--playground)
+- Data: `string` - HTML-based value of the input typed.
 - inputSettings:
 
 | Name | Type | Description |
 | ---- | ---- | ----------- |
-| **`direction`** | `"rtl" | "ltr" | ""` | optional - Defines the starting point at which the typed words will be displayed, "lrt" starts from the left moving to the right and "rlt" vice versa. Default is "rlt". |
-| **`language`** | `string` | optional - Defines the language in which the assistive elements of the text editor will be displayed. For example the placeholder and the character and word counter. |
-| **`maxCharacters`** | `number` | optional - Defines the limit of characters that are allow to type in the input element. |
-| **`spellcheck`** | `boolean` | optional - If it's enabled the text editor will mark misspellings. |
+| `controls` | `ControlsConfig` | optional - Defines the control configuration. See [controls](#controls) below for more information. |
+| `extensions` | `Extensions` | optional - The extensions to be loaded by the underlying Tiptap library. This property will override default extensions defined by Mosaic. Those default extensions are listed [here](https://github.com/simpleview/sv-mosaic/tree/develop/containers/mosaic/src/components/Field/FormFieldTextEditor/Extensions/defaultExtensions.ts). See the Tiptap [documentation](https://tiptap.dev/docs/editor/extensions/custom-extensions) for more information on creating extensions. |
+| `onImage` | `(params: TextEditorOnLinkParams) => void` | A function to be invoked when the image control is clicked. It will recieve a single parameter which is an object containing `updateImage`, `alt` and `src`. `updateImage` should be subsequently called to add/update the image. `alt` and `src` will correspond to the current image if applicable. |
+| `onLink` | `(params: TextEditorOnLinkParams) => void` | A function to be invoked when the link control is clicked. It will recieve a single parameter which is an object containing `updateLink`, `href` and `newTab`. `updateLink` should be subsequently called to add/update the link. `href` and `newTab` will correspond to the current image if applicable. |
+| `maxCharacters` | `number` | optional - Defines the limit of characters that are allow to type in the input element. |
 
-### How to use in a form?
+
+### Controls
+The `controls` input settings accepts a number of ways to define the controls that appear in the text editor's toolbar. The property is an array of groups, each group is an array of controls. Each item of the group can optionally be a third nested array which will produce a menu button which, when clicked, will reveal the controls listed within. By default, the controls configuration is as follows:
 
 ```ts
-const fields: FieldDef[] = useMemo(
-	() =>
-		[
-			//...other fields
-			{
-				//...all generic field props
-				type: "textEditor",
-				inputSettings: {
-					spellcheck: true,
-					direction: "rlt",
-					language: "en",
-					maxCharacters: 20,
-				},
+const defaultControls: ControlsConfig = [
+	// Control group containing a single control
+	["headings"],
+	// Control group containing 2 controls and a menu which contains 5 more controls.
+	["bold", "italic", ["underline", "strike", "superscript", "subscript", "clear"]],
+	["bulletList", "orderedList"],
+	["alignLeft", "alignCenter", ["alignRight", "alignJustify"]],
+	["link", ["image", "codeBlock", "blockquote"]],
+	["undo", "redo"],
+];
+```
+
+The control identifiers shown in the example above map to a control definition object, but you can also pass that definition object instead of the identifying string for more granular control. This example uses the `controlUndo` and `controlRedo` object available from Mosaic with a slight modification:
+
+```ts
+import { controlUndo, controlRedo } from "@sv-mosaic/mosaic/components/Field/FormFieldTextEditor";
+
+const controls: ControlsConfig = [
+	// Other controls...
+	[
+		{
+			...controlUndo,
+			label: "Roll it back",
+		},
+		{
+			...controlRedo,
+			label: "Restore it",
+		}
+	],
+];
+```
+
+But you could also define the entire control object to implement your own functionality:
+
+```ts
+import { controlUndo, controlRedo } from "@simpleview/mosaic/components/Field/FormFieldTextEditor";
+
+const controls: ControlsConfig = [
+	// Other controls...
+	[
+		{
+			name: "selectAll",
+			label: "Select All",
+			cmd: ({ editor }) => editor.chain().selectAll().run(),
+			Icon: SelectAllIcon,
+			shortcut: {
+				mac: "Cmd+A",
+				pc: "Ctrl+A",
 			},
-			//...other fields
-		],
-	[]
-);
+		}
+	],
+];
 ```
